Update system to use absolute rankings
Cards are now ranked by absolute rankings on a scale of 0.0 to 5.0.
Relative ranking was removed because it unnecessarily constrained the
ability to add user input in the future. Conceptually it is much easier to
update a card's absolute ranking than to rerank it relative to all the
other cards in the set.
Add a script to parse LSV articles. It is not 100% perfect, due to cycles
being condensed into one review, and inconsistent HTML.